Tc#83 1. All the logic for tracking the current logo and sawed into the hook. 2. There is a display of the current logo in the footer
