+++ frontend/src/shared/components/shared/footer/footer.tsx
@@ -1,6 +1,8 @@
+'use client';
+
 import { DropdownMenu, DropdownMenuContent, DropdownMenuItem, DropdownMenuTrigger } from '@/shared/components/ui';
 import { API_URL, CURRENT_YEAR_FULL, NAVIGATION_ITEMS } from '@/shared/constants';
-import { cn, useAppSelector } from '@/shared/lib';
+import {cn, useAppSelector} from '@/shared/lib';
 import { selectUserPermission } from '@/shared/lib/features/users/users-slice';
 import { FooterElementsData } from '@/shared/types/footer.types';
 import { ChevronUpIcon } from '@radix-ui/react-icons';
@@ -10,6 +12,7 @@ import React from 'react';
 import { SocialIcon } from 'react-social-icons';
 
 import styles from './footer.module.css';
+import {useNavbarLogo} from '@/shared/components/shared/navbar/use-navbar-logo';
 
 interface Props {
   dataItems: FooterElementsData[];
@@ -17,13 +20,15 @@ interface Props {
 
 export const Footer: React.FC<Props> = ({ dataItems }) => {
   const userPermission = useAppSelector(selectUserPermission);
+  const { currentLogo } = useNavbarLogo();
+
   return (
     <div className={cn(styles.footer, 'dark:bg-gray-900')}>
       <div className={styles.footerWrapper}>
         <div className={styles.footerColumn}>
           <div className={styles.logoWrapper}>
             <Link prefetch={true} href='/' className={styles.logo}>
-              <img src='/kslt.svg' alt='КСЛТ' />
+              <img className={styles.logo} src={API_URL + '/' + (currentLogo || '/kslt.svg')} alt='КСЛТ'/>
             </Link>
           </div>
           <div className={styles.copyright}>
